Changed to seperate foot in single.html
Facebook Pixel script forces itself onto every page and loads inside the body instead of the head. In an attempt to isotale the code, a partial has made for the Pixel code loads only on head2. Script seems to load on every page even after the code being absent from other pages. Trying to use a separate footer for single.html to see if it stops the behaviour.