Update privacy page to have a grey background only
The original background image was stretching and pixelating. In order to load a different css for that page alone, some conditional logic was used in the head to load a different css file, which imports the main file, and just adds code to hide the image.

A bit complex, but best implementation when using a static site. Also easy to maintain and adjust in the future.