Added Netlify Identity widget
Commented out in footer. Styling and sizing is ready to go. Just need to remove comments. As for page redirect, Netlify provides event handlers for login/logout, etc. in the widget, like:

`netlifyIdentity.on('login', user => console.log(user))`

That one will console log the user object when the `user` logs in. There are multiple event handlers available as part of the `netlifyIdentity` object. You can redirect using `window.location`  in that object. Netlify doesn't provide the code for redirecting, but the events are there and you can redirect in them as I described.